SDKC done
+The SDKCheck functionality is implemented. In the AppSCD.json, the
developer can now specify a list of frameworks, (that are embedded in
the app bundle) for which the access to some specified input sources
shall be denied. This means that -void methods that are called will not
have any effect and methods which return a result return a "falsy" one,
i.e zero, NO, nil and empty collection data structures.